# monolith/static/swagger.json
{
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation of auto trade API.\n\nSchemes:\nhttp\nhttps",
    "title": "infoblog.",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/api/1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Авторизация пользователя.",
        "operationId": "loginRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/loginResponse"
          }
        }
      }
    },
    "/api/1/auth/refresh": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "auth"
        ],
        "summary": "Обновление рефреш токена.",
        "operationId": "refreshRequest",
        "responses": {
          "200": {
            "$ref": "#/responses/refreshResponse"
          }
        }
      }
    },
    "/api/1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Регистрация пользователя.",
        "operationId": "registerRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/registerResponse"
          }
        }
      }
    },
    "/api/1/auth/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Верификация почты/телефона пользователя.",
        "operationId": "verifyRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/VerifyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/verifyResponse"
          }
        }
      }
    },
...
    "/api/1/bot/toggle": {
      "post": {
        "tags": [
          "bot"
        ],
        "summary": "Активация, деактивация бота.",
        "operationId": "botToggleRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/BotToggleRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/botToggleResponse"
          }
        }
      }
    },
...
    "/api/1/strategy/create": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "strategy"
        ],
        "summary": "Создание стратегии.",
        "operationId": "strategyCreateRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StrategyCreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/strategyCreateResponse"
          },
          "400": {
            "description": " Bad request"
          },
          "500": {
            "description": " Internal server error"
          }
        }
      }
    },
    "/api/1/strategy/list": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "strategy"
        ],
        "summary": "Получение списка стратегий.",
        "operationId": "strategyGetListRequest",
        "responses": {
          "200": {
            "$ref": "#/responses/strategyGetListResponse"
          },
          "400": {
            "description": " Bad request"
          },
          "500": {
            "description": " Internal server error"
          }
        }
      }
    },
    "/api/1/strategy/update": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "strategy"
        ],
        "summary": "Редактирование стратегии.",
        "operationId": "strategyUpdateRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StrategyUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/strategyUpdateResponse"
          },
          "400": {
            "description": " Bad request"
          },
          "500": {
            "description": " Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AuthResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/LoginData"
        },
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/auth/controller"
    },
...
    "Bot": {
      "description": "go:generate easytags $GOFILE json,mapper",
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "asset_type": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AssetType"
        },
        "auto_limit_sell_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "AutoLimitSellPercent"
        },
        "auto_sell": {
          "type": "boolean",
          "x-go-name": "AutoSell"
        },
        "commission_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "CommissionPercent"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "deleted_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DeletedAt"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "exchange_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExchangeID"
        },
        "exchange_type": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExchangeType"
        },
        "exchange_user_key_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExchangeUserKeyID"
        },
        "fixed_amount": {
          "type": "number",
          "format": "double",
          "x-go-name": "FixedAmount"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "kind": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Kind"
        },
        "limit_buy_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "LimitBuyPercent"
        },
        "limit_order": {
          "type": "boolean",
          "x-go-name": "LimitOrder"
        },
        "limit_sell_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "LimitSellPercent"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "order_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "OrderCount"
        },
        "order_count_limit": {
          "type": "boolean",
          "x-go-name": "OrderCountLimit"
        },
        "order_type": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "OrderType"
        },
        "pair_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "PairID"
        },
        "pairs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StrategyPair"
          },
          "x-go-name": "Pairs"
        },
        "sell_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "SellPercent"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedAt"
        },
        "user_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        },
        "uuid": {
          "type": "string",
          "x-go-name": "UUID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/models"
    },
...
    "BotToggleRequest": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "bot_uuid": {
          "type": "string",
          "x-go-name": "BotUUID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/bot/controller"
    },
...
    "Data": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/user/controller"
    },
    "DefaultResponse": {
      "type": "object",
      "properties": {
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/bot/controller"
    },
...
    "LoginData": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string",
          "x-go-name": "AccessToken"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "refresh_token": {
          "type": "string",
          "x-go-name": "RefreshToken"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/auth/controller"
    },
    "LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/auth/controller"
    },
...
    "RegisterRequest": {
      "description": "go:generate easytags $GOFILE",
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "idempotency_key": {
          "type": "string",
          "x-go-name": "IdempotencyKey"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "retype_password": {
          "type": "string",
          "x-go-name": "RetypePassword"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/auth/controller"
    },
    "RegisterResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/Data"
        },
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/auth/controller"
    },
...
    "StrategiesResponse": {
      "type": "object",
      "properties": {
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "strategy": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Strategy"
          },
          "x-go-name": "Strategy"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/strategy/controller"
    },
    "Strategy": {
      "description": "go:generate easytags $GOFILE json,mapper",
      "type": "object",
      "properties": {
        "bots": {
          "description": "UserID      int    `json:\"user_id\" mapper:\"user_id\"`\nAPIKey      string `json:\"api_key\" mapper:\"api_key\"`  TODO: нужно ли?\nSecretKey   string `json:\"secret_key\" mapper:\"secret_key\"`",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Bot"
          },
          "x-go-name": "Bots"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "exchange_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExchangeID"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "uuid": {
          "type": "string",
          "x-go-name": "UUID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/models"
    },
    "StrategyCreateRequest": {
      "type": "object",
      "properties": {
        "bots": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Bot"
          },
          "x-go-name": "Bots"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "exchange_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExchangeID"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "uuid": {
          "type": "string",
          "x-go-name": "UUID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/strategy/controller"
    },
    "StrategyCreateResponse": {
      "type": "object",
      "properties": {
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "strategy_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StrategyID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/strategy/controller"
    },
...
    "StrategyPair": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "pair_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "PairID"
        },
        "strategy_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StrategyID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/models"
    },
...
    "StrategyUpdateRequest": {
      "type": "object",
      "properties": {
        "fields": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "Fields"
        },
        "strategy": {
          "$ref": "#/definitions/Strategy"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/strategy/controller"
    },
...
    "User": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "email_verified": {
          "type": "boolean",
          "x-go-name": "EmailVerified"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "phone": {
          "type": "string",
          "x-go-name": "Phone"
        },
        "phone_verified": {
          "type": "boolean",
          "x-go-name": "PhoneVerified"
        },
        "role": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Role"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        },
        "verified": {
          "type": "boolean",
          "x-go-name": "Verified"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/models"
    },
...
    "VerifyRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "hash": {
          "type": "string",
          "x-go-name": "Hash"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/auth/controller"
    },
...
  },
  "responses": {
...
    "botToggleResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/DefaultResponse"
      }
    },
...
    "loginResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/AuthResponse"
      }
    },
...
    "refreshResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/AuthResponse"
      }
    },
    "registerResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/RegisterResponse"
      }
    },
    "strategyCreateResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/StrategyCreateResponse"
      }
    },
...
    "strategyGetListResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/StrategiesResponse"
      }
    },
...
    "verifyResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/AuthResponse"
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}